% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aire.zmvm-deprecated.R
\name{aire.zmvm-deprecated}
\alias{aire.zmvm-deprecated}
\alias{get_latest_data}
\alias{get_zone_data}
\title{Deprecated functions in package \pkg{aire.zmvm}.}
\usage{
get_latest_data(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{Details}{

\tabular{rl}{
  \code{get_zone_data} \tab now a synonym for \code{\link{get_zone_imeca}}\cr
  \code{get_latest_data} \tab now a synonym for \code{\link{get_latest_imeca}}\cr
}
}

\keyword{internal}
