% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aire.zmvm-package.R
\docType{package}
\name{aire.zmvm-package}
\alias{aire.zmvm}
\alias{aire.zmvm-package}
\title{\code{aire.zmvm} package}
\description{
Tools for downloading data from each of the pollution, wind and temperature
measuring stations in the Zona Metropolitana del Valle de México
(greater Mexico City).
}
\details{
See the README on
\href{https://github.com/diegovalle/aire.zmvm#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hoyodesmog.diegovalle.net/aire.zmvm/}
  \item \url{https://github.com/diegovalle/aire.zmvm}
  \item Report bugs at \url{https://github.com/diegovalle/aire.zmvm/issues}
}

}
\author{
\strong{Maintainer}: Diego Valle-Jones \email{diego@diegovalle.net}

}
\keyword{internal}
