% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aire.zmvm.R
\docType{package}
\name{aire.zmvm}
\alias{aire.zmvm}
\alias{aire.zmvm-package}
\title{\code{aire.zmvm} package}
\description{
Tools for downloading data from each of the pollution, wind and temperature
measuring stations in the Zona Metropolitana del Valle de México
(greater Mexico City).
}
\details{
See the README on
\href{https://github.com/diegovalle/aire.zmvm#readme}{GitHub}
}
