% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_imeca}
\alias{convert_to_imeca}
\title{Convert pollution values to IMECA}
\usage{
convert_to_imeca(value, pollutant, showWarnings = TRUE)
}
\arguments{
\item{value}{a numeric vector of values to convert to IMECAs}

\item{pollutant}{type of pollutant. A vector of one or more of the
following options:
\itemize{
 \item{"SO2"}{ - Dioxido de azufre}
 \item{"CO"}{ - Monoxido de carbono}
 \item{"NO2"}{ - Dioxido de nitrogeno}
 \item{"O3"}{ - Ozono}
 \item{"PM10"}{ - Particulas menores a 10 micrometros}
}}

\item{showWarnings}{deprecated; you can use the function
\code{\link[base]{suppressWarnings}} instead.}
}
\value{
value in IMECAs
}
\description{
Air quality in Mexico City is reported in IMECAs. This function converts
pollution values in the original units returned by sensors to
\href{https://en.wikipedia.org/wiki/Índice_Metropolitano_de_la_Calidad_del_Aire}{IMECA}
(Indice Metropolitano de la Calidad del Aire), a dimensionless scale
where all the pollutants can be compared.
}
\examples{
## IMECAs are a dimensionless scale that allows for the comparison of
## different pollutants
convert_to_imeca(157, "O3")
convert_to_imeca(c(450, 350, 250), rep("NO2", 3))
convert_to_imeca(80, "PM10")

## warning about recycling elements in a vector
convert_to_imeca(c(157, 200), "O3")

## suppress all warnings
suppressWarnings(convert_to_imeca(c(157, 200), "O3"))

convert_to_imeca(67, "O3")
convert_to_imeca(77, "O3")
convert_to_imeca(205, "O3")
convert_to_imeca(72, "O3")
convert_to_imeca(98, "O3")

}
\seealso{
\url{http://rama.edomex.gob.mx/contaminacion-atmosferica/imeca}
}
