% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aire.zmvm-deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{get_latest_data}
\alias{get_zone_data}
\alias{get_station_single_month}
\title{Deprecated functions in package \pkg{aire.zmvm}.}
\usage{
get_zone_data(...)
get_latest_data(...)
get_station_single_month(...)

get_latest_data(...)

get_zone_data(...)

get_station_single_month(...)
}
\arguments{
\item{...}{Parameters to be passed to the modern version of the function}
}
\value{
A data.frame

A data.frame

A data.frame

A data.frame
}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned.
}
\section{Details}{

\tabular{rl}{
  \code{get_zone_data} \tab now a synonym for
  \code{\link{get_zone_imeca}}\cr
  \code{get_latest_data} \tab now a synonym for
  \code{\link{get_latest_imeca}}\cr
  \code{get_station_single_month} \tab now a synonym for
  \code{\link{get_station_month_data}(criterion = "HORARIOS", ...)}\cr
}
}

\keyword{internal}
