% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_archives.R
\name{.clean_archive}
\alias{.clean_archive}
\title{Clean archive files}
\usage{
.clean_archive(df, include_hour)
}
\arguments{
\item{df}{data.frame to clean}

\item{include_hour}{is the data hourly}
}
\value{
data.frame
}
\description{
Clean archive files
}
\keyword{internal}
