% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_imeca.R
\name{.convert_time}
\alias{.convert_time}
\title{Convert a the time string from the aire.cdmx website to a date}
\usage{
.convert_time(time_div)
}
\arguments{
\item{time_div}{time to convert}
}
\description{
Convert a the time string from the aire.cdmx website to a date
}
\keyword{internal}
