% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_data.R
\name{.download_current_station_data}
\alias{.download_current_station_data}
\title{Title}
\usage{
.download_current_station_data(criterion, pollutant, year, month = "")
}
\arguments{
\item{criterion}{type of data to download}

\item{pollutant}{type of pollutant to download}

\item{year}{year to download}

\item{month}{month to download}
}
\description{
Title
}
\keyword{internal}
