% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zone_imeca.R
\name{.download_data_zone}
\alias{.download_data_zone}
\title{Title}
\usage{
.download_data_zone(criterion, pollutant, zone, start_date, end_date)
}
\arguments{
\item{criterion}{bla}

\item{pollutant}{bla}

\item{zone}{bla}

\item{start_date}{bla}

\item{end_date}{bla}
}
\value{
data.frame
}
\description{
Title
}
\keyword{internal}
