% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_data.R
\name{.download_old_station_data}
\alias{.download_old_station_data}
\title{Title}
\usage{
.download_old_station_data(pollutant, year)
}
\arguments{
\item{pollutant}{pollutant to filter by}

\item{year}{year to download}
}
\description{
Title
}
\keyword{internal}
