% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_archive_wsp}
\alias{.get_archive_wsp}
\title{Hack to download 2016 WSP HORARIO data since the csv files were converted
from mph to m/s (when the original data were already in m/s)}
\usage{
.get_archive_wsp(syear)
}
\arguments{
\item{syear}{shortyear in YY format}
}
\value{
data.frame
}
\description{
Hack to download 2016 WSP HORARIO data since the csv files were converted
from mph to m/s (when the original data were already in m/s)
}
\keyword{internal}
