% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recode_pollutant}
\alias{.recode_pollutant}
\title{Recode pollutant abbreviations}
\usage{
.recode_pollutant(pollutant)
}
\arguments{
\item{pollutant}{type of pollutant (O3, SO2, etc)}
}
\value{
data.frame
}
\description{
Recode pollutant abbreviations
}
\keyword{internal}
