% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recode_unit}
\alias{.recode_unit}
\title{Recode numeric codes to concentration units}
\usage{
.recode_unit(pollutant)
}
\arguments{
\item{pollutant}{type of pollutant (O3, SO2, etc)}
}
\value{
data.frame
}
\description{
Recode numeric codes to concentration units
}
\keyword{internal}
