% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recode_unit_code}
\alias{.recode_unit_code}
\title{Recode numeric codes to concentration units (extended)}
\usage{
.recode_unit_code(code)
}
\arguments{
\item{code}{numeric code for the the units}
}
\value{
data.frame
}
\description{
Recode numeric codes to concentration units (extended)
}
\keyword{internal}
