% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_archives.R
\name{download_meteorological}
\alias{download_meteorological}
\title{Download Meteorological Data Archives}
\usage{
download_meteorological(year, progress = interactive())
}
\arguments{
\item{year}{a numeric vector containing the years for which to download data
(the earliest possible value is 1986)}

\item{progress}{whether to display a progress bar (TRUE or FALSE).
By default it will only display in an interactive session.}
}
\value{
a data.frame with meterological information:
"RH","TMP","WDR","WSP","PBa"
}
\description{
Download the files available at
\href{http://www.aire.cdmx.gob.mx/default.php?opc=\%27aKBhnmI=\%27&opcion=Zw==}{Meteorología}
}
\examples{
\dontrun{
head(download_meteorological(2017))
}
}
