% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_archives.R
\name{download_pollution}
\alias{download_pollution}
\title{Download Pollution Archives}
\usage{
download_pollution(year, progress = interactive())
}
\arguments{
\item{year}{a numeric vector containing the years for which to download data
(the earliest possible value is 2009)}

\item{progress}{whether to display a progress bar (TRUE or FALSE).
By default it will only display in an interactive session.}
}
\value{
a data.frame with pollution information for the following pollutants
"CO", "NO", "NO2", "NOX", "O3", "PM10", "SO2", "PM25", and "PMCO"
}
\description{
Download the pollution files available at
\href{http://www.aire.cdmx.gob.mx/default.php?opc=\%27aKBhnmI=\%27&opcion=Zg==}{Contaminante}
}
\examples{
\dontrun{
head(download_pollution(2017))
}
}
