% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_archives.R
\name{download_pressure}
\alias{download_pressure}
\title{Download Atmospheric Pressure Archives}
\usage{
download_pressure(year, progress = interactive())
}
\arguments{
\item{year}{a numeric vector containing the years for which to download data
(the earliest possible value is 2009)}

\item{progress}{whether to display a progress bar (TRUE or FALSE).
By default it will only display in an interactive session.}
}
\value{
A data.frame with atmospheric pressure data.
}
\description{
The data comes from \href{http://www.aire.cdmx.gob.mx/default.php?opc=\%27aKBhnmI=\%27&opcion=bg==}{Presión Atmosférica}
}
\examples{
\dontrun{
head(download_pressure(2017))
}
}
