% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zone_data.R
\name{download_zone}
\alias{download_zone}
\title{Title}
\usage{
download_zone(criterion, pollutant, zone, start_date, end_date)
}
\arguments{
\item{criterion}{bla}

\item{pollutant}{bla}

\item{zone}{bla}

\item{start_date}{bla}

\item{end_date}{bla}
}
\value{
data.frame
}
\description{
Title
}
