% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_data.R
\name{get_latest_data}
\alias{get_latest_data}
\title{Get the latest pollution values for each station}
\usage{
get_latest_data()
}
\value{
A data.frame with pollution values in IMECAs, the time corresponds to the
America/Mexico_City timezone
}
\description{
Get the latest pollution values for each station
}
\examples{
df <- get_latest_data()
head(df)
}
