% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution.R
\name{get_pollution_data}
\alias{get_pollution_data}
\title{Download pollution data from the aire.df.gob.mx server.
http://www.aire.df.gob.mx/estadisticas-consultas/concentraciones/index.php}
\usage{
get_pollution_data(tipo, parametro, anio)
}
\arguments{
\item{tipo}{Type of data to download. Can be "HORARIOS" for hourly averages, "MAXIMOS" for the
daily maximum and "MINIMOS" for the daily minimum}

\item{parametro}{The type of pollutant to download. Can be of type "so2", "co", "nox", "no2",
"no", "o3", "pm10", "pm2", "wsp", "wdr", "tmp", "rh"}

\item{anio}{The years to download}
}
\value{
data.frame with pollution data
}
\description{
Download pollution data from the aire.df.gob.mx server.
http://www.aire.df.gob.mx/estadisticas-consultas/concentraciones/index.php
}
\examples{
\dontrun{
#Download daily maximum pm10 data from 2005 to 2016
df <- get_pollution_data("MAXIMOS", "pm10", 2005:2016)
head(df)
}
}

