% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_data.R
\name{get_station_data}
\alias{get_station_data}
\title{Download pollution data}
\usage{
get_station_data(criterion, pollutant, year)
}
\arguments{
\item{criterion}{Type of data to download.
\itemize{
 \item{"HORARIOS"}{ - Hourly data}
 \item{"MAXIMOS""}{ - Daily maximums}
 \item{"MINIMOS"}{ - Daily minimums}
}}

\item{pollutant}{The type of pollutant to download.
\itemize{
 \item{"SO2"}{ - Dioxido de azufre}
 \item{"CO"}{ - Monoxido de carbono}
 \item{"NOX"}{ - Oxidos de nitrogeno}
 \item{"NO2"}{ - Dioxido de nitrogeno}
 \item{"NO"}{ - Oxido nitrico}
 \item{"O3"}{ - Ozono}
 \item{"PM10"}{ - Particulas menores a 10 micrometros}
 \item{"PM2"}{ - Particulas menores a 2.5 micrometros}
 \item{"WSP"}{ - Velocidad del viento}
 \item{"WDR"}{ - Direccion del viento}
 \item{"TMP"}{ - Temperatura ambiente}
 \item{"RH"}{ - Humedad relativa}
}}

\item{year}{a numeric vector containing the years for which to download data}
}
\value{
a data.frame with pollution data
}
\description{
retrieve pollution data from the air quality server at \url{
http://www.aire.df.gob.mx/estadisticas-consultas/concentraciones/index.php}
}
\examples{
\dontrun{
# Download daily maximum PM10 data (particulate matter 10 micrometers or less in diameter)
# from 2015 to 2016
df <- get_station_data("MAXIMOS", "pm10", 2015:2016)
head(df)
}
}

