% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zone_imeca.R
\name{get_zone_imeca}
\alias{get_zone_imeca}
\title{Download pollution data by zone in IMECAs}
\usage{
get_zone_imeca(criterion, pollutant, zone, start_date, end_date,
  showWarnings = TRUE, show_messages = TRUE)
}
\arguments{
\item{criterion}{The type of data to download
\itemize{
 \item{"HORARIOS"}{ - Hourly data}
 \item{"MAXIMOS""}{ - Daily maximums}
}}

\item{pollutant}{The type of pollutant to download
\itemize{
 \item{"SO2"}{ - Dioxido de azufre}
 \item{"CO"}{ - Monoxido de carbono}
 \item{"NO2"}{ - Dioxido de nitrogeno}
 \item{"O3"}{ - Ozono}
 \item{"PM10"}{ - Particulas menores a 10 micrometros}
 \item{"TC"}{- All the pollutants}
}}

\item{zone}{The geographic zone for which to download data
\itemize{
 \item{"NO"}{ - Noroeste}
 \item{"NE"}{ - Noreste}
 \item{"CE"}{ - Centro}
 \item{"SO"}{ - Suroeste}
 \item{"SE"}{ - Sureste}
 \item{"TZ"}{ - All zones}
}}

\item{start_date}{The start date in YYYY-MM-DD format (earliest possible value is 2008-01-01).}

\item{end_date}{The end date in YYYY-MM-DD format.}

\item{showWarnings}{Deprecated; use show_messages instead.}

\item{show_messages}{show a message about issues with performing the conversion}
}
\value{
A data.frame with pollution data measured in IMECAS, by geographic zone. The hours
correspond to the \emph{Etc/GMT+6} timezone, with no daylight saving time
}
\description{
retrieve pollution data in IMECAs by geographic zone from the air quality
server at \url{http://www.aire.cdmx.gob.mx/default.php?opc='aqBjnmU='}
}
\details{
Note that the which stations belong to which zones may change over time, also
the way the IMECA is computed can change so that comparison of
pollution values across time may not be valid. In
2015 it was determined that the stations with codes ACO, AJU, INN, MON
and MPA would no longer be taken into consideration when computing the
pollution index and at some point in the future would be no longer
included in the data by zone.

The different geographic zones were defined in the Gaceta Oficial
CDMX, No 106, 1 de julio 2016.

\strong{Zona Centro}: Benito Juárez,
Cuauhtémoc, Iztacalco and Venustiano Carranza.

\strong{Zona Noreste}: Gustavo A. Madero, Coacalco de Berriozábal,
Chicoloapan, Chimalhuacán,
Ecatepec de Morelos, Ixtapaluca, La Paz,
Nezahualcóyotl and Tecámac.

\strong{Zona Noroeste}: Azcapotzalco,
Miguel Hidalgo, Atizapán de Zaragoza, Cuautitlán, Cuautitlán Izcalli,
Naucalpan de Juárez, Nicolás
Romero, Tlalnepantla de Baz and Tultitlán.

\strong{Zona Sureste}: Iztapalapa, Milpa
Alta, Tláhuac, Xochimilco, Chalco and Valle de Chalco.

\strong{Zona Suroeste}: Álvaro Obregón,
Coyoacán, Cuajimalpa, Magdalena Contreras, Tlalpan and Huixquilucan.
}
\examples{
## There was a regional (NE) PM10 pollution emergency on Jan 6, 2017
get_zone_imeca("MAXIMOS", "PM10", "NE", "2017-01-05", "2017-01-08",
               show_messages = FALSE)

## There was an ozone pollution emergency on May 15, 2017
get_zone_imeca("MAXIMOS", "O3", "TZ", "2017-05-15", "2017-05-15",
               show_messages = FALSE)

\dontrun{
## Download daily maximum PM10 data (particulate matter 10 micrometers or
## less in diameter) from 2015-01-01 to 2016-03-20 for all geographic zones
df <- get_zone_imeca("MAXIMOS", "PM10", "TZ", "2015-01-01", "2016-03-20")
head(df)

## Download hourly O3 pollution data for May 15, 2017. Only the suroeste zone
df2 <- get_zone_imeca("HORARIOS", "O3", "SO", "2017-05-15", "2017-05-15")

## Convert to local Mexico City time
df2$mxc_time <- format(as.POSIXct(paste0(df2$date, " ", df2$hour, ":00"),
                                  tz = "Etc/GMT+6"),
                       tz = "America/Mexico_City")

head(df2)
}

}
