% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Pollution measuring stations in Mexico City}
\format{A data frame with 63 rows and 7 variables:
\describe{
  \item{cve_estac}{abbreviation of the station}
  \item{nom_estac}{name of the station}
  \item{longitud}{longitud of the station}
  \item{latitud}{latitude of the station}
  \item{alt}{altitude of the station}
  \item{obs_estac}{comment}
  \item{id_station}{id of the station}

}}
\source{
\url{http://148.243.232.112:8080/opendata/catalogos/cat_estacion.csv}
}
\usage{
stations
}
\description{
Pollution measuring stations in Mexico City
}
\keyword{datasets}

