% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_imeca}
\alias{convert_to_imeca}
\title{Title}
\usage{
convert_to_imeca(pollutant, value)
}
\arguments{
\item{pollutant}{type of pollutant}

\item{value}{value to convert to IMECAS}
}
\value{
value in IMECAS
}
\description{
Title
}

