% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_imeca.R
\name{get_latest_imeca}
\alias{get_latest_imeca}
\title{Get the latest pollution values for each station}
\usage{
get_latest_imeca()
}
\value{
A data.frame with pollution values in IMECAs, the hour corresponds
to the \emph{America/Mexico_City} timezone
}
\description{
Note that in
2015 it was determined that the stations with codes ACO, AJU, INN, MON
and MPA would no longer be taken into consideration when computing the
pollution index and are not included in the data returned by this function
}
\examples{
df <- get_latest_imeca()
head(df)
}
