% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution.R
\name{get_pollution_data}
\alias{get_pollution_data}
\title{Download pollution data}
\usage{
get_pollution_data(tipo, parametro, anio)
}
\arguments{
\item{tipo}{Type of data to download. Can be "HORARIOS" for hourly averages, "MAXIMOS" for the
daily maximum and "MINIMOS" for the daily minimum}

\item{parametro}{The type of pollutant to download. Can be of type "so2", "co", "nox", "no2",
"no", "o3", "pm10", "pm2", "wsp", "wdr", "tmp", "rh"}

\item{anio}{a vector containing the years for which to download data}
}
\value{
a data.frame with pollution data
}
\description{
retrieve pollution data from the air quality server at \url{
http://www.aire.df.gob.mx/estadisticas-consultas/concentraciones/index.php}
}
\examples{
\dontrun{
# Download daily maximum PM10 data (particulate matter 10 micrometers or less in diameter)
# from 2005 to 2016
df <- get_pollution_data("MAXIMOS", "pm10", 2005:2016)
head(df)
}
}

