% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_imeca.R
\name{get_station_imeca}
\alias{get_station_imeca}
\title{Download pollution data by station in IMECAS}
\usage{
get_station_imeca(pollutant, date, show_messages = TRUE)
}
\arguments{
\item{pollutant}{The type of pollutant to download
\itemize{
 \item{"SO2"}{ - Dioxido de azufre}
 \item{"CO"}{ - Monoxido de carbono}
 \item{"NO2"}{ - Dioxido de nitrogeno}
 \item{"O3"}{ - Ozono}
 \item{"PM10"}{ - Particulas menores a 10 micrometros}
}}

\item{date}{The date for which to download data in YYYY-MM-DD format
(the earliest possible date is 2009-01-01).}

\item{show_messages}{show a message about issues with performing the
conversion}
}
\value{
A data.frame with pollution data measured in IMECAS, by station.
The hours correspond to the \emph{Etc/GMT+6} timezone, with no daylight
saving time
}
\description{
Note that in
2015 it was determined that the stations with codes ACO, AJU, INN, MON
and MPA would no longer be taken into consideration when computing the
pollution index and at some point in the future would no longer be inclued
in the data returned by this function
}
\examples{
## There was an ozone pollution emergency on May 15, 2017
df_o3 <- get_station_imeca("O3", "2017-05-15", show_messages = FALSE)

## Convert to local Mexico City time
df_o3$mxc_time <- format(as.POSIXct(paste0(df_o3$date,
                                           " ",
                                           df_o3$hour,
                                           ":00"),
                                    tz = "Etc/GMT+6"),
                         tz = "America/Mexico_City")
head(df_o3[order(-df_o3$value), ])
}
