% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw360.R
\name{idw360}
\alias{idw360}
\title{Inverse Distance Weighting with Directional Data}
\usage{
idw360(values, coords, grid, idp = 2)
}
\arguments{
\item{values}{the dependent variable}

\item{coords}{the spatial data locations where the values were measured}

\item{grid}{data frame or Spatial object with the locations to predict}

\item{idp}{The inverse distance weighting power}
}
\value{
data.frame with the interpolated values for each of the grid points
}
\description{
function for inverse distance weighted interpolation with directional data. Useful for when you
are working with data whose unit of measurement is degrees (i.e. the average of 35° and 355°
should be 15°). It works by finding the shortest distance between two degree marks
on a circle.
}
\examples{
library(sp)
# Location of sensors. First column x/longitud, second y/latitude
locations <- data.frame(lon = c(1, 2), lat = c(1, 2))
coordinates(locations) <- ~lon+lat
# Could be wind direction values in degrees
values <- c(55, 355)
# The grid for which to extrapolate values
grid <- data.frame(lon = c(1, 2, 1, 2), lat = c(1, 2, 2, 1))
coordinates(grid) <- ~lon+lat
idw360(values, locations, grid)

}
